[2.15 - 2.19*] Phonebook: backend integration + toasts
- Save new contacts
- Delete/update existing contacts

(*) I had already implemented error notificaitions
using react-toastify. I understand pretty well the
solution provided in the notes and I would be surely
able to implement it without even looking at the notes
again, but I think it is not very reusable and elegant, so I'm not going to remove react-toastify for that.

Because exercise 2.19 is in the "Adding styles"
section and I've already added a lot more styles
that what required for 2.19, I think I can consider
2.19 solved even though I didn't implement the
"notification" part of the exercise.